[export] Separate out graph signature
[ghstack-poisoned]
